agriculture-rag/examples/requirements.txt
sentence-transformers
requests
msgpack
numpy
google-genai